backend/requirements.txt
<<<<<<< HEAD
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.19
httpx==0.28.1
cryptography==43.0.1
boto3==1.35.78
python-multipart==0.0.17
orjson==3.10.15
pytesseract==0.3.13
Pillow==11.1.0
=======
fastapi==0.111.0
uvicorn[standard]==0.30.1
strawberry-graphql==0.209.0
celery==5.3.6
redis==5.0.4
httpx==0.27.0
SQLAlchemy==2.0.29
psycopg[binary]==3.1.18
python-multipart==0.0.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
>>>>>>> main
